requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==3.2.15
django-appconf==1.0.5
django-compressor==4.1
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
factory-boy==3.2.1
Faker==14.2.0
phonenumbers==8.12.54
Pillow==9.2.0
purl==1.6
pycountry==22.3.5
python-dateutil==2.8.2
pytz==2022.2.1
rcssmin==1.1.0
rjsmin==1.2.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
tzdata==2022.2
